LayoutTests/imported/w3c/web-platform-tests/css/css-values/calc-infinity-nan-serialize-angle-expected.txt


FAIL 'rotate(calc(1deg * NaN))' as a specified value should serialize as 'rotate(calc(NaN * 1deg))'. assert_equals: 'rotate(calc(NaN * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(NaN * 1deg))" but got "rotate(calc(NaNdeg))"
FAIL 'rotate(calc(1rad * NaN))' as a specified value should serialize as 'rotate(calc(NaN * 1rad))'. assert_equals: 'rotate(calc(NaN * 1rad))' should round-trip exactly in specified values. expected "rotate(calc(NaN * 1rad))" but got "rotate(calc(NaNdeg))"
FAIL 'rotate(calc(1turn * NaN))' as a specified value should serialize as 'rotate(calc(NaN * 1turn))'. assert_equals: 'rotate(calc(NaN * 1turn))' should round-trip exactly in specified values. expected "rotate(calc(NaN * 1turn))" but got "rotate(calc(NaNdeg))"
FAIL 'rotate(calc(1grad * nan))' as a specified value should serialize as 'rotate(calc(NaN * 1grad))'. assert_equals: 'rotate(calc(NaN * 1grad))' should round-trip exactly in specified values. expected "rotate(calc(NaN * 1grad))" but got "rotate(calc(NaNdeg))"
FAIL 'rotate(calc(1deg * infinity / infinity))' as a specified value should serialize as 'rotate(calc(NaN * 1deg))'. assert_equals: 'rotate(calc(NaN * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(NaN * 1deg))" but got "rotate(calc(NaNdeg))"
FAIL 'rotate(calc(1deg * 0 * infinity))' as a specified value should serialize as 'rotate(calc(NaN * 1deg))'. assert_equals: 'rotate(calc(NaN * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(NaN * 1deg))" but got "rotate(calc(NaNdeg))"
FAIL 'rotate(calc(1deg * (infinity + -infinity)))' as a specified value should serialize as 'rotate(calc(NaN * 1deg))'. assert_equals: 'rotate(calc(NaN * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(NaN * 1deg))" but got "rotate(calc(NaNdeg))"
FAIL 'rotate(calc(1deg * (-infinity + infinity)))' as a specified value should serialize as 'rotate(calc(NaN * 1deg))'. assert_equals: 'rotate(calc(NaN * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(NaN * 1deg))" but got "rotate(calc(NaNdeg))"
FAIL 'rotate(calc(1deg * (infinity - infinity)))' as a specified value should serialize as 'rotate(calc(NaN * 1deg))'. assert_equals: 'rotate(calc(NaN * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(NaN * 1deg))" but got "rotate(calc(NaNdeg))"
FAIL 'rotate(calc(1deg * infinity))' as a specified value should serialize as 'rotate(calc(infinity * 1deg))'. assert_equals: 'rotate(calc(infinity * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(infinity * 1deg))" but got "rotate(calc(infinitydeg))"
FAIL 'rotate(calc(1deg * -infinity))' as a specified value should serialize as 'rotate(calc(-infinity * 1deg))'. assert_equals: 'rotate(calc(-infinity * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(-infinity * 1deg))" but got "rotate(calc(-infinitydeg))"
FAIL 'rotate(calc(1deg * iNFinIty))' as a specified value should serialize as 'rotate(calc(infinity * 1deg))'. assert_equals: 'rotate(calc(infinity * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(infinity * 1deg))" but got "rotate(calc(infinitydeg))"
FAIL 'rotate(calc(1deg * (infinity + infinity)))' as a specified value should serialize as 'rotate(calc(infinity * 1deg))'. assert_equals: 'rotate(calc(infinity * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(infinity * 1deg))" but got "rotate(calc(infinitydeg))"
FAIL 'rotate(calc(1deg * (-infinity + -infinity)))' as a specified value should serialize as 'rotate(calc(-infinity * 1deg))'. assert_equals: 'rotate(calc(-infinity * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(-infinity * 1deg))" but got "rotate(calc(-infinitydeg))"
PASS 'rotate(calc(1deg * 1/infinity))' as a specified value should serialize as 'rotate(calc(0deg))'.
FAIL 'rotate(calc(1deg * infinity * infinity))' as a specified value should serialize as 'rotate(calc(infinity * 1deg))'. assert_equals: 'rotate(calc(infinity * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(infinity * 1deg))" but got "rotate(calc(infinitydeg))"
FAIL 'rotate(calc(1deg * -infinity * -infinity))' as a specified value should serialize as 'rotate(calc(infinity * 1deg))'. assert_equals: 'rotate(calc(infinity * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(infinity * 1deg))" but got "rotate(calc(infinitydeg))"
FAIL 'rotate(calc(1 * max(INFinity*3deg, 0deg)))' as a specified value should serialize as 'rotate(calc(infinity * 1deg))'. assert_equals: 'rotate(calc(infinity * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(infinity * 1deg))" but got "rotate(calc(infinitydeg))"
PASS 'rotate(calc(1 * min(inFInity*4deg, 0deg)))' as a specified value should serialize as 'rotate(calc(0deg))'.
FAIL 'rotate(calc(1 * max(nAn*2deg, 0deg)))' as a specified value should serialize as 'rotate(calc(NaN * 1deg))'. assert_equals: 'rotate(calc(NaN * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(NaN * 1deg))" but got "rotate(calc(NaNdeg))"
FAIL 'rotate(calc(1 * min(nan*3deg, 0deg)))' as a specified value should serialize as 'rotate(calc(NaN * 1deg))'. assert_equals: 'rotate(calc(NaN * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(NaN * 1deg))" but got "rotate(calc(NaNdeg))"
FAIL 'rotate(calc(1 * clamp(-INFINITY*20deg, 0deg, infiniTY*10deg)))' as a specified value should serialize as 'rotate(calc(0deg))'. assert_equals: 'rotate(calc(1 * clamp(-INFINITY*20deg, 0deg, infiniTY*10deg)))' and 'rotate(calc(0deg))' should serialize the same in specified values. expected "rotate(calc(0deg))" but got "rotate(calc(1 * clamp(-infinitydeg, 0deg, infinitydeg)))"
FAIL 'rotate(calc(1deg * max(NaN, min(0,10))))' as a specified value should serialize as 'rotate(calc(NaN * 1deg))'. assert_equals: 'rotate(calc(NaN * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(NaN * 1deg))" but got "rotate(calc(NaNdeg))"
FAIL 'rotate(calc(1deg * clamp(NaN, 0, 10)))' as a specified value should serialize as 'rotate(calc(NaN * 1deg))'. assert_equals: 'rotate(calc(NaN * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(NaN * 1deg))" but got "rotate(calc(NaNdeg))"
FAIL 'rotate(calc(1deg * max(0, min(10, NaN))))' as a specified value should serialize as 'rotate(calc(NaN * 1deg))'. assert_equals: 'rotate(calc(NaN * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(NaN * 1deg))" but got "rotate(calc(NaNdeg))"
FAIL 'rotate(calc(1deg * clamp(0, 10, NaN)))' as a specified value should serialize as 'rotate(calc(NaN * 1deg))'. assert_equals: 'rotate(calc(NaN * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(NaN * 1deg))" but got "rotate(calc(NaNdeg))"
FAIL 'rotate(calc(1deg * max(0, min(NaN, 10))))' as a specified value should serialize as 'rotate(calc(NaN * 1deg))'. assert_equals: 'rotate(calc(NaN * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(NaN * 1deg))" but got "rotate(calc(NaNdeg))"
FAIL 'rotate(calc(1deg * clamp(0, NaN, 10)))' as a specified value should serialize as 'rotate(calc(NaN * 1deg))'. assert_equals: 'rotate(calc(NaN * 1deg))' should round-trip exactly in specified values. expected "rotate(calc(NaN * 1deg))" but got "rotate(calc(NaNdeg))"
FAIL 'rotate(calc(1deg * clamp(-Infinity, 0, infinity)))' as a specified value should serialize as 'rotate(calc(0deg))'. assert_equals: 'rotate(calc(1deg * clamp(-Infinity, 0, infinity)))' and 'rotate(calc(0deg))' should serialize the same in specified values. expected "rotate(calc(0deg))" but got "rotate(calc(1deg * clamp(-infinity, 0, infinity)))"
FAIL 'rotate(calc(1deg * clamp(-inFinity, infinity, 10)))' as a specified value should serialize as 'rotate(calc(10deg))'. assert_equals: 'rotate(calc(1deg * clamp(-inFinity, infinity, 10)))' and 'rotate(calc(10deg))' should serialize the same in specified values. expected "rotate(calc(10deg))" but got "rotate(calc(1deg * clamp(-infinity, infinity, 10)))"
